refactorize dtos [refactorize]
