+++ src/main/java/io/github/matheusfy/screanmatch/model/dtos/TemporadaDTO.java
@@ -1,4 +1,4 @@
-package io.github.matheusfy.screanmatch.model;
+package io.github.matheusfy.screanmatch.model.dtos;
 
 import com.fasterxml.jackson.annotation.JsonAlias;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
@@ -6,9 +6,9 @@ import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import java.util.List;
 
 @JsonIgnoreProperties(ignoreUnknown = true)
-public record DadosTemporadaDTO(
+public record TemporadaDTO(
         @JsonAlias("Season")  Integer numero,
-        @JsonAlias("Episodes")  List<DadosEpisodioDTO> lstEpisodios
+        @JsonAlias("Episodes")  List<EpisodioDTO> lstEpisodios
 
 ) {
 }
